// src/functions-api-ts-client/models/FunctionJob.ts
/**
 * Swagger Functions Store - OpenAPI 3.0
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { JobStatus } from '../models/JobStatus';
import { HttpFile } from '../http/http';

export class FunctionJob {
    'id'?: number | null;
    'functionID': number;
    'status'?: JobStatus | null;
    'inputs'?: any | null;
    'outputs'?: any | null;
    'jobInfo'?: any | null;
    'createdAt'?: Date | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "functionID",
            "baseName": "functionID",
            "type": "number",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "JobStatus",
            "format": ""
        },
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "any",
            "format": ""
        },
        {
            "name": "outputs",
            "baseName": "outputs",
            "type": "any",
            "format": ""
        },
        {
            "name": "jobInfo",
            "baseName": "job_info",
            "type": "any",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return FunctionJob.attributeTypeMap;
    }

    public constructor() {
    }
}