// src/functions-api-ts-client/types/ObjectParamAPI.ts
import { ResponseContext, RequestContext, HttpFile, HttpInfo } from '../http/http';
import { Configuration, ConfigurationOptions } from '../configuration'
import type { Middleware } from '../middleware';

import { Function } from '../models/Function';
import { FunctionJob } from '../models/FunctionJob';
import { FunctionJobCollection } from '../models/FunctionJobCollection';
import { HTTPValidationError } from '../models/HTTPValidationError';
import { JobStatus } from '../models/JobStatus';
import { ValidationError } from '../models/ValidationError';
import { ValidationErrorLocInner } from '../models/ValidationErrorLocInner';

import { ObservableDefaultApi } from "./ObservableAPI";
import { DefaultApiRequestFactory, DefaultApiResponseProcessor} from "../apis/DefaultApi";

export interface DefaultApiGenerateOpenapiGenerateOpenapiGetRequest {
}

export class ObjectDefaultApi {
    private api: ObservableDefaultApi

    public constructor(configuration: Configuration, requestFactory?: DefaultApiRequestFactory, responseProcessor?: DefaultApiResponseProcessor) {
        this.api = new ObservableDefaultApi(configuration, requestFactory, responseProcessor);
    }

    /**
     * Generate OpenAPI spec and save to file
     * Generate Openapi
     * @param param the request object
     */
    public generateOpenapiGenerateOpenapiGetWithHttpInfo(param: DefaultApiGenerateOpenapiGenerateOpenapiGetRequest = {}, options?: ConfigurationOptions): Promise<HttpInfo<any>> {
        return this.api.generateOpenapiGenerateOpenapiGetWithHttpInfo( options).toPromise();
    }

    /**
     * Generate OpenAPI spec and save to file
     * Generate Openapi
     * @param param the request object
     */
    public generateOpenapiGenerateOpenapiGet(param: DefaultApiGenerateOpenapiGenerateOpenapiGetRequest = {}, options?: ConfigurationOptions): Promise<any> {
        return this.api.generateOpenapiGenerateOpenapiGet( options).toPromise();
    }

}

import { ObservableFunctionApi } from "./ObservableAPI";
import { FunctionApiRequestFactory, FunctionApiResponseProcessor} from "../apis/FunctionApi";

export interface FunctionApiBatchRunFunctionRequest {
    /**
     * 
     * Defaults to: undefined
     * @type number
     * @memberof FunctionApibatchRunFunction
     */
    functionId: number
    /**
     * 
     * Defaults to: undefined
     * @type string
     * @memberof FunctionApibatchRunFunction
     */
    collectionName: string
    /**
     * 
     * @type Array&lt;string&gt;
     * @memberof FunctionApibatchRunFunction
     */
    requestBody: Array<string>
    /**
     * 
     * Defaults to: undefined
     * @type number
     * @memberof FunctionApibatchRunFunction
     */
    maxWorkers?: number
}

export interface FunctionApiCreateFunctionRequest {
    /**
     * 
     * @type Function
     * @memberof FunctionApicreateFunction
     */
    _function: Function
}

export interface FunctionApiDeleteAllFunctionsRequest {
}

export interface FunctionApiListFunctionsRequest {
}

export interface FunctionApiMapFunctionRequest {
    /**
     * 
     * Defaults to: undefined
     * @type number
     * @memberof FunctionApimapFunction
     */
    functionId: number
    /**
     * 
     * @type Array&lt;string&gt;
     * @memberof FunctionApimapFunction
     */
    requestBody: Array<string>
    /**
     * 
     * Defaults to: undefined
     * @type number
     * @memberof FunctionApimapFunction
     */
    maxWorkers?: number
}

export interface FunctionApiRunFunctionRequest {
    /**
     * 
     * Defaults to: undefined
     * @type number
     * @memberof FunctionApirunFunction
     */
    functionId: number
    /**
     * 
     * Defaults to: undefined
     * @type string
     * @memberof FunctionApirunFunction
     */
    inputs: string
}

export interface FunctionApiSearchFunctionsByNameRequest {
    /**
     * 
     * Defaults to: undefined
     * @type string
     * @memberof FunctionApisearchFunctionsByName
     */
    name: string
}

export interface FunctionApiSearchFunctionsByTagsRequest {
    /**
     * Tags to search for
     * Defaults to: undefined
     * @type Array&lt;string&gt;
     * @memberof FunctionApisearchFunctionsByTags
     */
    tags: Array<string>
    /**
     * If True, functions must have all tags. If False, functions must have any of the tags.
     * Defaults to: false
     * @type boolean
     * @memberof FunctionApisearchFunctionsByTags
     */
    matchAll?: boolean
}

export interface FunctionApiUpdateFunctionConfigFunctionConfigPostRequest {
    /**
     * 
     * Defaults to: 10
     * @type number
     * @memberof FunctionApiupdateFunctionConfigFunctionConfigPost
     */
    maxParallelJobs?: number
}

export class ObjectFunctionApi {
    private api: ObservableFunctionApi

    public constructor(configuration: Configuration, requestFactory?: FunctionApiRequestFactory, responseProcessor?: FunctionApiResponseProcessor) {
        this.api = new ObservableFunctionApi(configuration, requestFactory, responseProcessor);
    }

    /**
     * Run a function with multiple inputs and create a job collection.  Parameters:     function_id: ID of the function to run     collection_name: Name for the job collection     request_body: List of JSON strings containing input parameters     max_workers: Optional maximum number of parallel workers  Returns:     Created function job collection containing all job IDs
     * Batch Run Function
     * @param param the request object
     */
    public batchRunFunctionWithHttpInfo(param: FunctionApiBatchRunFunctionRequest, options?: ConfigurationOptions): Promise<HttpInfo<FunctionJobCollection>> {
        return this.api.batchRunFunctionWithHttpInfo(param.functionId, param.collectionName, param.requestBody, param.maxWorkers,  options).toPromise();
    }

    /**
     * Run a function with multiple inputs and create a job collection.  Parameters:     function_id: ID of the function to run     collection_name: Name for the job collection     request_body: List of JSON strings containing input parameters     max_workers: Optional maximum number of parallel workers  Returns:     Created function job collection containing all job IDs
     * Batch Run Function
     * @param param the request object
     */
    public batchRunFunction(param: FunctionApiBatchRunFunctionRequest, options?: ConfigurationOptions): Promise<FunctionJobCollection> {
        return this.api.batchRunFunction(param.functionId, param.collectionName, param.requestBody, param.maxWorkers,  options).toPromise();
    }

    /**
     * Create a new function with optional JSON Schema definitions for input and output. Validates that provided schemas are valid JSON Schema. Supports tags for better organization and searchability.
     * Create Function
     * @param param the request object
     */
    public createFunctionWithHttpInfo(param: FunctionApiCreateFunctionRequest, options?: ConfigurationOptions): Promise<HttpInfo<Function>> {
        return this.api.createFunctionWithHttpInfo(param._function,  options).toPromise();
    }

    /**
     * Create a new function with optional JSON Schema definitions for input and output. Validates that provided schemas are valid JSON Schema. Supports tags for better organization and searchability.
     * Create Function
     * @param param the request object
     */
    public createFunction(param: FunctionApiCreateFunctionRequest, options?: ConfigurationOptions): Promise<Function> {
        return this.api.createFunction(param._function,  options).toPromise();
    }

    /**
     * Delete all functions from the store.  Returns:     Message confirming deletion with count of deleted functions
     * Delete All Functions
     * @param param the request object
     */
    public deleteAllFunctionsWithHttpInfo(param: FunctionApiDeleteAllFunctionsRequest = {}, options?: ConfigurationOptions): Promise<HttpInfo<any>> {
        return this.api.deleteAllFunctionsWithHttpInfo( options).toPromise();
    }

    /**
     * Delete all functions from the store.  Returns:     Message confirming deletion with count of deleted functions
     * Delete All Functions
     * @param param the request object
     */
    public deleteAllFunctions(param: FunctionApiDeleteAllFunctionsRequest = {}, options?: ConfigurationOptions): Promise<any> {
        return this.api.deleteAllFunctions( options).toPromise();
    }

    /**
     * List all functions in the store.  Returns:     List of all registered functions
     * List Functions
     * @param param the request object
     */
    public listFunctionsWithHttpInfo(param: FunctionApiListFunctionsRequest = {}, options?: ConfigurationOptions): Promise<HttpInfo<Array<Function>>> {
        return this.api.listFunctionsWithHttpInfo( options).toPromise();
    }

    /**
     * List all functions in the store.  Returns:     List of all registered functions
     * List Functions
     * @param param the request object
     */
    public listFunctions(param: FunctionApiListFunctionsRequest = {}, options?: ConfigurationOptions): Promise<Array<Function>> {
        return this.api.listFunctions( options).toPromise();
    }

    /**
     * Start asynchronous processing of multiple inputs with schema validation.
     * Map Function
     * @param param the request object
     */
    public mapFunctionWithHttpInfo(param: FunctionApiMapFunctionRequest, options?: ConfigurationOptions): Promise<HttpInfo<Array<FunctionJob>>> {
        return this.api.mapFunctionWithHttpInfo(param.functionId, param.requestBody, param.maxWorkers,  options).toPromise();
    }

    /**
     * Start asynchronous processing of multiple inputs with schema validation.
     * Map Function
     * @param param the request object
     */
    public mapFunction(param: FunctionApiMapFunctionRequest, options?: ConfigurationOptions): Promise<Array<FunctionJob>> {
        return this.api.mapFunction(param.functionId, param.requestBody, param.maxWorkers,  options).toPromise();
    }

    /**
     * Run a function with the given inputs. Validates inputs and outputs against JSON Schema if defined.
     * Run Function
     * @param param the request object
     */
    public runFunctionWithHttpInfo(param: FunctionApiRunFunctionRequest, options?: ConfigurationOptions): Promise<HttpInfo<FunctionJob>> {
        return this.api.runFunctionWithHttpInfo(param.functionId, param.inputs,  options).toPromise();
    }

    /**
     * Run a function with the given inputs. Validates inputs and outputs against JSON Schema if defined.
     * Run Function
     * @param param the request object
     */
    public runFunction(param: FunctionApiRunFunctionRequest, options?: ConfigurationOptions): Promise<FunctionJob> {
        return this.api.runFunction(param.functionId, param.inputs,  options).toPromise();
    }

    /**
     * Search for functions by name.  Parameters:     name: String to search for in function names (case-sensitive partial match)  Returns:     List of functions whose names contain the search string
     * Search Functions By Name
     * @param param the request object
     */
    public searchFunctionsByNameWithHttpInfo(param: FunctionApiSearchFunctionsByNameRequest, options?: ConfigurationOptions): Promise<HttpInfo<Array<Function>>> {
        return this.api.searchFunctionsByNameWithHttpInfo(param.name,  options).toPromise();
    }

    /**
     * Search for functions by name.  Parameters:     name: String to search for in function names (case-sensitive partial match)  Returns:     List of functions whose names contain the search string
     * Search Functions By Name
     * @param param the request object
     */
    public searchFunctionsByName(param: FunctionApiSearchFunctionsByNameRequest, options?: ConfigurationOptions): Promise<Array<Function>> {
        return this.api.searchFunctionsByName(param.name,  options).toPromise();
    }

    /**
     * Search for functions by tags.  Parameters:     tags: List of tags to search for     match_all: If True, functions must have all specified tags. If False, functions must have any of the specified tags.  Returns:     List of functions that match the tag criteria
     * Search Functions By Tags
     * @param param the request object
     */
    public searchFunctionsByTagsWithHttpInfo(param: FunctionApiSearchFunctionsByTagsRequest, options?: ConfigurationOptions): Promise<HttpInfo<Array<Function>>> {
        return this.api.searchFunctionsByTagsWithHttpInfo(param.tags, param.matchAll,  options).toPromise();
    }

    /**
     * Search for functions by tags.  Parameters:     tags: List of tags to search for     match_all: If True, functions must have all specified tags. If False, functions must have any of the specified tags.  Returns:     List of functions that match the tag criteria
     * Search Functions By Tags
     * @param param the request object
     */
    public searchFunctionsByTags(param: FunctionApiSearchFunctionsByTagsRequest, options?: ConfigurationOptions): Promise<Array<Function>> {
        return this.api.searchFunctionsByTags(param.tags, param.matchAll,  options).toPromise();
    }

    /**
     * Update function execution configuration settings.  Parameters:     max_parallel_jobs: Maximum number of parallel jobs allowed (default: 10)  Returns:     Updated configuration settings
     * Update Function Config
     * @param param the request object
     */
    public updateFunctionConfigFunctionConfigPostWithHttpInfo(param: FunctionApiUpdateFunctionConfigFunctionConfigPostRequest = {}, options?: ConfigurationOptions): Promise<HttpInfo<any>> {
        return this.api.updateFunctionConfigFunctionConfigPostWithHttpInfo(param.maxParallelJobs,  options).toPromise();
    }

    /**
     * Update function execution configuration settings.  Parameters:     max_parallel_jobs: Maximum number of parallel jobs allowed (default: 10)  Returns:     Updated configuration settings
     * Update Function Config
     * @param param the request object
     */
    public updateFunctionConfigFunctionConfigPost(param: FunctionApiUpdateFunctionConfigFunctionConfigPostRequest = {}, options?: ConfigurationOptions): Promise<any> {
        return this.api.updateFunctionConfigFunctionConfigPost(param.maxParallelJobs,  options).toPromise();
    }

}

import { ObservableFunctionJobApi } from "./ObservableAPI";
import { FunctionJobApiRequestFactory, FunctionJobApiResponseProcessor} from "../apis/FunctionJobApi";

export interface FunctionJobApiGetFunctionJobRequest {
    /**
     * 
     * Defaults to: undefined
     * @type number
     * @memberof FunctionJobApigetFunctionJob
     */
    functionJobId: number
}

export interface FunctionJobApiGetFunctionJobsRequest {
    /**
     * ID of the function to get jobs for
     * Defaults to: undefined
     * @type number
     * @memberof FunctionJobApigetFunctionJobs
     */
    functionId: number
    /**
     * Maximum number of jobs to return
     * Minimum: 1
     * Maximum: 1000
     * Defaults to: undefined
     * @type number
     * @memberof FunctionJobApigetFunctionJobs
     */
    limit?: number
    /**
     * Number of jobs to skip
     * Minimum: 0
     * Defaults to: undefined
     * @type number
     * @memberof FunctionJobApigetFunctionJobs
     */
    offset?: number
    /**
     * Filter by job status
     * Defaults to: undefined
     * @type JobStatus
     * @memberof FunctionJobApigetFunctionJobs
     */
    status?: JobStatus
    /**
     * Filter jobs after this date
     * Defaults to: undefined
     * @type Date
     * @memberof FunctionJobApigetFunctionJobs
     */
    startDate?: Date
    /**
     * Filter jobs before this date
     * Defaults to: undefined
     * @type Date
     * @memberof FunctionJobApigetFunctionJobs
     */
    endDate?: Date
}

export interface FunctionJobApiGetJobsStatusRequest {
    /**
     * 
     * Defaults to: undefined
     * @type Array&lt;number&gt;
     * @memberof FunctionJobApigetJobsStatus
     */
    jobIds: Array<number>
}

export interface FunctionJobApiListFunctionJobsRequest {
    /**
     * Maximum number of jobs to return
     * Minimum: 1
     * Maximum: 1000
     * Defaults to: undefined
     * @type number
     * @memberof FunctionJobApilistFunctionJobs
     */
    limit?: number
    /**
     * Number of jobs to skip
     * Minimum: 0
     * Defaults to: undefined
     * @type number
     * @memberof FunctionJobApilistFunctionJobs
     */
    offset?: number
    /**
     * Filter by job status
     * Defaults to: undefined
     * @type JobStatus
     * @memberof FunctionJobApilistFunctionJobs
     */
    status?: JobStatus
    /**
     * Filter by function ID
     * Defaults to: undefined
     * @type number
     * @memberof FunctionJobApilistFunctionJobs
     */
    functionId?: number
    /**
     * Filter jobs after this date
     * Defaults to: undefined
     * @type Date
     * @memberof FunctionJobApilistFunctionJobs
     */
    startDate?: Date
    /**
     * Filter jobs before this date
     * Defaults to: undefined
     * @type Date
     * @memberof FunctionJobApilistFunctionJobs
     */
    endDate?: Date
}

export class ObjectFunctionJobApi {
    private api: ObservableFunctionJobApi

    public constructor(configuration: Configuration, requestFactory?: FunctionJobApiRequestFactory, responseProcessor?: FunctionJobApiResponseProcessor) {
        this.api = new ObservableFunctionJobApi(configuration, requestFactory, responseProcessor);
    }

    /**
     * Get the details of a specific function job.  Parameters:     function_job_id: ID of the function job to retrieve  Returns:     Function job details including status, inputs, and outputs  Raises:     HTTPException: If job is not found (404)
     * Get Function Job
     * @param param the request object
     */
    public getFunctionJobWithHttpInfo(param: FunctionJobApiGetFunctionJobRequest, options?: ConfigurationOptions): Promise<HttpInfo<FunctionJob>> {
        return this.api.getFunctionJobWithHttpInfo(param.functionJobId,  options).toPromise();
    }

    /**
     * Get the details of a specific function job.  Parameters:     function_job_id: ID of the function job to retrieve  Returns:     Function job details including status, inputs, and outputs  Raises:     HTTPException: If job is not found (404)
     * Get Function Job
     * @param param the request object
     */
    public getFunctionJob(param: FunctionJobApiGetFunctionJobRequest, options?: ConfigurationOptions): Promise<FunctionJob> {
        return this.api.getFunctionJob(param.functionJobId,  options).toPromise();
    }

    /**
     * Get all jobs for a specific function with optional filtering and pagination.
     * Get Function Jobs
     * @param param the request object
     */
    public getFunctionJobsWithHttpInfo(param: FunctionJobApiGetFunctionJobsRequest, options?: ConfigurationOptions): Promise<HttpInfo<Array<FunctionJob>>> {
        return this.api.getFunctionJobsWithHttpInfo(param.functionId, param.limit, param.offset, param.status, param.startDate, param.endDate,  options).toPromise();
    }

    /**
     * Get all jobs for a specific function with optional filtering and pagination.
     * Get Function Jobs
     * @param param the request object
     */
    public getFunctionJobs(param: FunctionJobApiGetFunctionJobsRequest, options?: ConfigurationOptions): Promise<Array<FunctionJob>> {
        return this.api.getFunctionJobs(param.functionId, param.limit, param.offset, param.status, param.startDate, param.endDate,  options).toPromise();
    }

    /**
     * Get status of multiple jobs.  Parameters:     job_ids: List of job IDs to check  Returns:     List of job statuses
     * Get Jobs Status
     * @param param the request object
     */
    public getJobsStatusWithHttpInfo(param: FunctionJobApiGetJobsStatusRequest, options?: ConfigurationOptions): Promise<HttpInfo<Array<FunctionJob>>> {
        return this.api.getJobsStatusWithHttpInfo(param.jobIds,  options).toPromise();
    }

    /**
     * Get status of multiple jobs.  Parameters:     job_ids: List of job IDs to check  Returns:     List of job statuses
     * Get Jobs Status
     * @param param the request object
     */
    public getJobsStatus(param: FunctionJobApiGetJobsStatusRequest, options?: ConfigurationOptions): Promise<Array<FunctionJob>> {
        return this.api.getJobsStatus(param.jobIds,  options).toPromise();
    }

    /**
     * List all function jobs with optional filtering and pagination.  Parameters:     limit: Maximum number of jobs to return (default: all)     offset: Number of jobs to skip for pagination (default: 0)     status: Filter by job status (e.g., PENDING, RUNNING, COMPLETED, FAILED)     function_id: Filter jobs for a specific function     start_date: Include jobs created after this date     end_date: Include jobs created before this date  Returns:     List[FunctionJob]: A filtered list of function jobs
     * List all function jobs with optional filtering
     * @param param the request object
     */
    public listFunctionJobsWithHttpInfo(param: FunctionJobApiListFunctionJobsRequest = {}, options?: ConfigurationOptions): Promise<HttpInfo<Array<FunctionJob>>> {
        return this.api.listFunctionJobsWithHttpInfo(param.limit, param.offset, param.status, param.functionId, param.startDate, param.endDate,  options).toPromise();
    }

    /**
     * List all function jobs with optional filtering and pagination.  Parameters:     limit: Maximum number of jobs to return (default: all)     offset: Number of jobs to skip for pagination (default: 0)     status: Filter by job status (e.g., PENDING, RUNNING, COMPLETED, FAILED)     function_id: Filter jobs for a specific function     start_date: Include jobs created after this date     end_date: Include jobs created before this date  Returns:     List[FunctionJob]: A filtered list of function jobs
     * List all function jobs with optional filtering
     * @param param the request object
     */
    public listFunctionJobs(param: FunctionJobApiListFunctionJobsRequest = {}, options?: ConfigurationOptions): Promise<Array<FunctionJob>> {
        return this.api.listFunctionJobs(param.limit, param.offset, param.status, param.functionId, param.startDate, param.endDate,  options).toPromise();
    }

}

import { ObservableFunctionJobCollectionApi } from "./ObservableAPI";
import { FunctionJobCollectionApiRequestFactory, FunctionJobCollectionApiResponseProcessor} from "../apis/FunctionJobCollectionApi";

export interface FunctionJobCollectionApiCreateFunctionJobCollectionRequest {
    /**
     * 
     * @type FunctionJobCollection
     * @memberof FunctionJobCollectionApicreateFunctionJobCollection
     */
    functionJobCollection: FunctionJobCollection
}

export interface FunctionJobCollectionApiGetCollectionStatusRequest {
    /**
     * 
     * Defaults to: undefined
     * @type number
     * @memberof FunctionJobCollectionApigetCollectionStatus
     */
    collectionId: number
}

export interface FunctionJobCollectionApiListFunctionJobCollectionsRequest {
}

export class ObjectFunctionJobCollectionApi {
    private api: ObservableFunctionJobCollectionApi

    public constructor(configuration: Configuration, requestFactory?: FunctionJobCollectionApiRequestFactory, responseProcessor?: FunctionJobCollectionApiResponseProcessor) {
        this.api = new ObservableFunctionJobCollectionApi(configuration, requestFactory, responseProcessor);
    }

    /**
     * Create a new function job collection.  Parameters:     collection: Collection details including name and optional description  Returns:     Created function job collection
     * Create Function Job Collection
     * @param param the request object
     */
    public createFunctionJobCollectionWithHttpInfo(param: FunctionJobCollectionApiCreateFunctionJobCollectionRequest, options?: ConfigurationOptions): Promise<HttpInfo<FunctionJobCollection>> {
        return this.api.createFunctionJobCollectionWithHttpInfo(param.functionJobCollection,  options).toPromise();
    }

    /**
     * Create a new function job collection.  Parameters:     collection: Collection details including name and optional description  Returns:     Created function job collection
     * Create Function Job Collection
     * @param param the request object
     */
    public createFunctionJobCollection(param: FunctionJobCollectionApiCreateFunctionJobCollectionRequest, options?: ConfigurationOptions): Promise<FunctionJobCollection> {
        return this.api.createFunctionJobCollection(param.functionJobCollection,  options).toPromise();
    }

    /**
     * Get status of a function job collection.  Parameters:     collection_id: ID of the collection to check  Returns:     Collection details including current status of all jobs
     * Get Collection Status
     * @param param the request object
     */
    public getCollectionStatusWithHttpInfo(param: FunctionJobCollectionApiGetCollectionStatusRequest, options?: ConfigurationOptions): Promise<HttpInfo<FunctionJobCollection>> {
        return this.api.getCollectionStatusWithHttpInfo(param.collectionId,  options).toPromise();
    }

    /**
     * Get status of a function job collection.  Parameters:     collection_id: ID of the collection to check  Returns:     Collection details including current status of all jobs
     * Get Collection Status
     * @param param the request object
     */
    public getCollectionStatus(param: FunctionJobCollectionApiGetCollectionStatusRequest, options?: ConfigurationOptions): Promise<FunctionJobCollection> {
        return this.api.getCollectionStatus(param.collectionId,  options).toPromise();
    }

    /**
     * List all function job collections.  Returns:     List of all function job collections
     * List Function Job Collections
     * @param param the request object
     */
    public listFunctionJobCollectionsWithHttpInfo(param: FunctionJobCollectionApiListFunctionJobCollectionsRequest = {}, options?: ConfigurationOptions): Promise<HttpInfo<Array<FunctionJobCollection>>> {
        return this.api.listFunctionJobCollectionsWithHttpInfo( options).toPromise();
    }

    /**
     * List all function job collections.  Returns:     List of all function job collections
     * List Function Job Collections
     * @param param the request object
     */
    public listFunctionJobCollections(param: FunctionJobCollectionApiListFunctionJobCollectionsRequest = {}, options?: ConfigurationOptions): Promise<Array<FunctionJobCollection>> {
        return this.api.listFunctionJobCollections( options).toPromise();
    }

}
